사용자별 적립금 사용내역 상세조회 평균 적립금 반올림 적용

+++ src/main/java/com/emoney/domain/dto/response/ResponseEmoneyUserDetailDto.java
@@ -14,6 +14,6 @@ public class ResponseEmoneyUserDetailDto {
     private Long userSeq;
     private Long usageCount;
     private String latestUsageDateTime;
-    private Double averageAmount;
+    private Long averageAmount;
     private Long totalAmount;
 }
